Extracted smartystreets logic into a custom hook
